Rewrite data fetching from scratch
Since the API changes a lot since this version
was written, I had to re write most of the
data fetching part. I also updated few types.
Usually I would separate it into a lot of smaller
commits, but it was hard to do in manner, that every
commit could work separately, so I just created big
one with all the cnahges.